phishing_detector/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.0
playwright>=1.37.0
opencv-python-headless>=4.8.0
imagehash>=4.3.1
Pillow>=10.0.0
scikit-image>=0.21.0
torch>=2.0.0
torchvision>=0.15.0
celery>=5.3.0
redis>=4.6.0
loguru>=0.7.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
scipy>=1.11.0
